Changes in progress to allow a cross-platform command-line version of NAADSM to interact with a database server for storing generated output.
Further changes may make it possible to run NAADSM as a multi-threaded application with Qt.  This ought to be relatively straightforward, and threads do not have to interact with one another.

The first step involved translating NAADSM output classes from Delphi to C++, and compiling the NAADSM core as a C++ application.  Full functionality of the new C++ data storage classes is not yet implemented, but code compiles cleanly.